Migrate codebase to TypeScript and update configurations (#4381)
* refactor: migrate JavaScript files to TypeScript and update configurations

* refactor: update import statements to use single quotes and enhance linting configurations

* refactor: update ESLint configuration to use TypeScript parser and convert PrismTheme to ES module export

* refactor: update lint:format script to include JSX and TypeScript file extensions

* refactor: update prettier script to include JSX and TypeScript file extensions

* refactor: update TypeScript configurations and improve type annotations in blog components

* refactor: remove broken hover effect from clean button in custom theme styles

* refactor: fix metadata usage in DocsRating and improve TypeScript types in various components

* refactor: adjust TypeScript type declaration for global `ga` function in DocsRating

* refactor: simplify Section component props using React.PropsWithChildren

* refactor: add TypeScript types for Google Analytics and clean up global declaration in DocsRating

* refactor: enhance TypeScript types for user authentication and streamline related components

* refactor: update Prettier commands in package.json for improved file matching

* refactor: update Prettier command in package.json for improved TypeScript file matching

* refactor: update Prettier commands in package.json for improved file matching

* refactor: update Prettier command for improved file matching and standardize package versions

* set json extension on prettier config file
