+++ plugins/remark-lint-no-dead-urls/package.json
@@ -14,7 +14,7 @@
     "src/*"
   ],
   "scripts": {
-    "prettier": "prettier --write \"{src/**/*.js,tests/**/*.js,*.md}\"",
+    "prettier": "prettier --write '{src,tests}/**/*.{md,js,jsx,ts,tsx}'",
     "test": "yarn node --experimental-vm-modules $(yarn bin jest)"
   },
   "dependencies": {
