+++ plugins/remark-snackplayer/package.json
@@ -15,7 +15,7 @@
     "src"
   ],
   "scripts": {
-    "prettier": "prettier --write \"{src/**/*.js,tests/**/*.js,*.md}\"",
+    "prettier": "prettier --write '{src,tests}/**/*.{md,js,jsx,ts,tsx}'",
     "test": "yarn tape tests/index.js"
   },
   "dependencies": {
