+++ website/package.json
@@ -14,23 +14,24 @@
     "start": "docusaurus start",
     "build": "docusaurus build && yarn run update-redirect ./build/_redirects ./versions.json",
     "build:fast": "PREVIEW_DEPLOY=true yarn run build",
+    "tsc": "npx tsc --noEmit",
     "swizzle": "docusaurus swizzle",
     "deploy": "docusaurus deploy",
     "serve": "docusaurus serve",
     "clear": "docusaurus clear",
     "test": "yarn build",
     "version:cut": "docusaurus docs:version",
-    "format:source": "prettier --write {{core,src}/**/*.js,*.js}",
+    "format:source": "prettier --write '{core,src}/**/*.{js,jsx,ts,tsx}'",
     "format:markdown": "prettier --write ../docs/*.md && prettier --write {versioned_docs/**/*.md,blog/*.md}",
     "format:style": "prettier --write src/**/*.{scss,css}",
     "format:examples": "eslint-examples-js --fix && eslint-examples-tsx --fix",
     "prettier": "yarn format:source && yarn format:markdown && yarn format:style",
-    "lint": "eslint ../docs/** blog/** core/** src/**/*.js ./*.js",
+    "lint": "eslint ../docs/** blog/** '{core,src}/**/*.{js,jsx,ts,tsx}'",
     "lint:examples": "eslint-examples-js && eslint-examples-tsx && tsc-examples",
     "lint:versioned": "eslint versioned_docs/**",
     "lint:markdown": "remark ../docs --quiet -r .remarkrc.mjs",
     "lint:markdown:versioned": "remark ./versioned_docs --quiet -r .remarkrc.mjs",
-    "lint:format": "prettier --check {{core,src}/**/*.js,*.js} ../docs/*.md {versioned_docs/**/*.md,blog/*.md} src/**/*.{scss,css}",
+    "lint:format": "prettier --check '{core,src}/**/*.{js,jsx,ts,tsx}' ../docs/*.md {versioned_docs/**/*.md,blog/*.md} src/**/*.{scss,css}",
     "language:lint": "cd ../ && alex && case-police 'docs/*.md' -p brands,general,products,softwares -d ./website/react-native-dict.json",
     "language:lint:versioned": "cd ../ && alex . && case-police '**/*.md' -p brands,general,products,softwares -d ./website/react-native-dict.json",
     "ci:lint": "yarn lint && yarn lint:examples && yarn lint:versioned && yarn language:lint:versioned && yarn lint:markdown && yarn lint:format",
@@ -49,11 +50,11 @@
     ]
   },
   "dependencies": {
-    "@docusaurus/core": "3.6.0",
-    "@docusaurus/faster": "3.6.0",
-    "@docusaurus/plugin-google-gtag": "3.6.0",
-    "@docusaurus/plugin-pwa": "3.6.0",
-    "@docusaurus/preset-classic": "3.6.0",
+    "@docusaurus/core": "3.6.3",
+    "@docusaurus/faster": "3.6.3",
+    "@docusaurus/plugin-google-gtag": "3.6.3",
+    "@docusaurus/plugin-pwa": "3.6.3",
+    "@docusaurus/preset-classic": "3.6.3",
     "docusaurus-plugin-sass": "^0.2.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -61,15 +62,19 @@
     "sass": "1.76.0"
   },
   "devDependencies": {
-    "@docusaurus/types": "3.6.0",
+    "@docusaurus/module-type-aliases": "3.6.3",
+    "@docusaurus/tsconfig": "3.6.3",
+    "@docusaurus/types": "3.6.3",
     "@react-native-website/lint-examples": "0.0.0",
     "@react-native-website/update-redirects": "0.0.0",
+    "@types/google.analytics": "^0.0.46",
     "alex": "^10.0.0",
     "fs-extra": "^10.1.0",
     "glob": "^8.0.3",
     "glob-promise": "^4.2.2",
     "path": "^0.12.7",
     "remark-cli": "^11.0.0",
-    "remark-lint-no-dead-urls": "^1.1.0"
+    "remark-lint-no-dead-urls": "^1.1.0",
+    "typescript": "^5.7.2"
   }
 }
