+++ website/src/css/customTheme.scss
@@ -697,10 +697,6 @@ a[class*="tagRegular"] {
     transform: rotate(180deg) scale(0.9);
   }
 
-  .clean-btn:hover {
-    background: var(--ifm-color-emphasis-800);
-  }
-
   .DocSearch-Button {
     border-radius: var(--ifm-global-radius);
     padding: 0 6px 0 10px;
