+++ website/src/pages/Home/Hero/FloorBackground.tsx
