+++ website/src/pages/Home/Platforms/FoxFact.tsx
