+++ website/src/pages/Home/components/SectionTitle/index.tsx
@@ -9,11 +9,19 @@ import React from 'react';
 
 import styles from './styles.module.css';
 
-function SectionTitle({title, description}) {
+function SectionTitle({
+  title,
+  description,
+}: {
+  title: string;
+  description?: React.ReactNode;
+}) {
   return (
     <div className={styles.container}>
       <h1 className={styles.title}>{title}</h1>
-      <h3 className={styles.description}>{description}</h3>
+      {description ? (
+        <h3 className={styles.description}>{description}</h3>
+      ) : null}
     </div>
   );
 }
