+++ website/src/pages/Home/components/ThemeImage/index.tsx
@@ -15,7 +15,8 @@ function ThemeImage({lightSrc, darkSrc, className, alt}) {
       mutations.forEach(mutation => {
         if (
           mutation.type === 'attributes' &&
-          mutation.attributeName === 'data-theme'
+          mutation.attributeName === 'data-theme' &&
+          mutation.target instanceof Element
         ) {
           setTheme(mutation.target.getAttribute('data-theme'));
         }
