+++ website/src/theme/Blog/Components/Author/index.tsx
@@ -4,6 +4,7 @@ import Link from '@docusaurus/Link';
 import AuthorSocials from '@theme/Blog/Components/Author/Socials';
 import Heading from '@theme/Heading';
 import styles from './styles.module.css';
+
 function MaybeLink(props) {
   if (props.href) {
     return <Link {...props} />;
