--- website/src/theme/BlogSidebar/Desktop/index.d.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-/// <reference types="react" />
-import type { Props } from "@theme/BlogSidebar/Desktop";
-export default function BlogSidebarDesktop({ sidebar }: Props): JSX.Element;
