--- website/src/theme/BlogSidebar/Mobile/index.d.ts
@@ -1,9 +1,0 @@
-/**
- * Copyright (c) Meta Platforms, Inc. and affiliates.
- *
- * This source code is licensed under the MIT license found in the
- * LICENSE file in the root directory of this source tree.
- */
-/// <reference types="react" />
-import type { Props } from "@theme/BlogSidebar/Mobile";
-export default function BlogSidebarMobile(props: Props): JSX.Element;
