+++ website/src/theme/DocVersionBanner/index.tsx
@@ -8,7 +8,10 @@ import {
   useDocVersionSuggestions,
 } from '@docusaurus/plugin-content-docs/client';
 import {ThemeClassNames} from '@docusaurus/theme-common';
-import {useDocsVersion, useDocsPreferredVersion} from '@docusaurus/plugin-content-docs/client';
+import {
+  useDocsVersion,
+  useDocsPreferredVersion,
+} from '@docusaurus/plugin-content-docs/client';
 
 function UnreleasedVersionLabel({siteTitle, versionMetadata}) {
   return (
