+++ website/versioned_docs/version-0.76/_fabric-native-components.jsx
@@ -1,6 +1,6 @@
-import React from "react";
-import IOSContent from "./fabric-native-components-ios.md";
-import AndroidContent from "./fabric-native-components-android.md";
+import React from 'react';
+import IOSContent from './fabric-native-components-ios.md';
+import AndroidContent from './fabric-native-components-android.md';
 
 export function FabricNativeComponentsIOS() {
   return <IOSContent />;
