+++ website/versioned_docs/version-0.76/_turbo-native-modules-components.jsx
@@ -1,6 +1,6 @@
-import React from "react";
-import IOSContent from "./turbo-native-modules-ios.md";
-import AndroidContent from "./turbo-native-modules-android.md";
+import React from 'react';
+import IOSContent from './turbo-native-modules-ios.md';
+import AndroidContent from './turbo-native-modules-android.md';
 
 export function TurboNativeModulesIOS() {
   return <IOSContent />;
